api/requirements.txt
python-etcd
etcd3
protobuf<=3.20.3
fastapi
uvicorn
mysql-connector-python
redis
pydantic
